chore: Update Docker installation instructions in README
The README file has been updated to provide clearer instructions for installing Jenkins using Docker. The changes include adding a step to pull the Docker image and updating the command to run the Docker container. This will make it easier for users to set up Jenkins using Docker.

Note: This commit message follows the convention of starting with a verb in the imperative form ("Update") and providing a concise summary of the changes made.